papers/nearest_centroids_merlin/requirements.txt
# Nearest Centroid Classification - Quantum ML
# Paper: Johri et al. (2020) - arXiv:2012.04145

cirq>=1.4.0
scikit-learn>=1.7.2
merlinquantum>=0.2.3
matplotlib>=3.8.0
torchvision>=0.20.0
seaborn>=0.13.0
pytest>=8.0.0